refactor: Move actions to modules directory
