+++ src/components/Toast/index.tsx
@@ -1,6 +1,5 @@
 import { Component } from './component';
 import { Icon } from './Icon';
-export { createToast, dismissToast } from '../Toast/actions';
 
 export const Toast = Component as typeof Component & { Icon: typeof Icon };
