+++ src/index.tsx
@@ -13,7 +13,7 @@ export {
 } from './modules/core';
 export { useBreadcrumbs, useAddBreadcrumbEffect, BreadcrumbProvider } from './modules/breadcrumbs';
 export { useBuildTestId, Testable } from './modules/test-ids';
-export { ToastProvider } from './modules/toast';
+export { createToast, dismissToast, ToastProvider } from './modules/toast';
 export { useMoonPayUrl, buildMoonPayUrl } from './modules/moonpay';
 export { formatCryptoAsset, formatFiatAsset } from './modules/intl';
 
@@ -53,6 +53,6 @@ export { Table } from './components/Table';
 export { Text } from './components/Text';
 export { TextArea } from './components/TextArea';
 export { TextInput } from './components/TextInput';
-export { createToast, dismissToast, Toast } from './components/Toast';
+export { Toast } from './components/Toast';
 export { Tooltip } from './components/Tooltip';
 export { Wallets, useWalletProviders } from './components/Wallets';
