+++ src/modules/toast/index.tsx
@@ -1,1 +1,2 @@
+export { createToast, dismissToast } from './actions';
 export { ToastProvider } from './ToastProvider';
